adding custom rpc from network settings
